travelock/requirements.txt
aniso8601==6.0.0
attrdict==2.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cytoolz==0.9.0.1
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.3
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.6.0
Flask==1.0.3
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
hexbytes==0.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lru-dict==1.1.6
MarkupSafe==1.1.1
marshmallow==2.19.2
parsimonious==0.8.1
pipenv==2018.11.26
py-solc==3.2.0
pycryptodome==3.8.2
pytz==2019.1
requests==2.22.0
rlp==1.1.0
semantic-version==2.6.0
six==1.12.0
toolz==0.9.0
urllib3==1.25.3
virtualenv==16.4.3
virtualenv-clone==0.5.3
web3==4.9.2
websockets==6.0
Werkzeug==0.15.4
